pyproject.toml
[project]
name = "faq"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jinja2>=3.1.6",
    "mistune>=3.0.0",
    "pygments>=2.19.2",
    "python-docx>=1.2.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.5",
    "minsearch>=0.0.7",
    "openai>=1.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
markers = [
    "unit: Unit tests for individual functions",
    "integration: Integration tests for complete workflows"
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]

[tool.setuptools]
packages = ["faq_automation"]
